+++ packages/init-testing/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @atomicsmash/init-testing
 
+## 1.1.0-beta.3
+
+### Patch Changes
+
+- [`cd493a3`](https://github.com/AtomicSmash/packages/commit/cd493a3eeea80e01ba1e693dcc621999ec3e55f6) Thanks [@mikeybinns](https://github.com/mikeybinns)! - Improve logging for package install complete and initialisers
+
+## 1.1.0-beta.2
+
+### Patch Changes
+
+- [`dc6ea3b`](https://github.com/AtomicSmash/packages/commit/dc6ea3bfe8631b76d12a1eb367596f8ede7be7c4) Thanks [@mikeybinns](https://github.com/mikeybinns)! - Add lighthouse folder to gitignore
+
+## 1.1.0-beta.1
+
+### Patch Changes
+
+- [`d384e18`](https://github.com/AtomicSmash/packages/commit/d384e182cd64886c786f36e079082a734c2e35e9) Thanks [@mikeybinns](https://github.com/mikeybinns)! - Run playwright installer after package install
+
+- [`71d0b8b`](https://github.com/AtomicSmash/packages/commit/71d0b8b1f631f71bc2b892e3e4958eda7e983dd3) Thanks [@mikeybinns](https://github.com/mikeybinns)! - Make ./playwright.config.ts work in commonjs projects
+
+## 1.1.0-beta.0
+
+### Minor Changes
+
+- [`aad23f8`](https://github.com/AtomicSmash/packages/commit/aad23f8bc1c200ea1dfbf6fc00d9a1673193fe3b) Thanks [@mikeybinns](https://github.com/mikeybinns)! - Check for site level ts for config
+
+### Patch Changes
+
+- [`e607650`](https://github.com/AtomicSmash/packages/commit/e607650e5e1afee161a1eeb4c026967695980be7) Thanks [@mikeybinns](https://github.com/mikeybinns)! - make files work in commonjs projects
+
 ## 1.0.0
 
 ### Major Changes